requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
gevent==1.2.2
google-api-python-client==1.6.4
google-auth==1.3.0
google-auth-httplib2==0.0.3
greenlet==0.4.12
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
msgpack-python==0.5.1
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-http-client==3.0.0
pyzmq==16.0.3
redis==2.10.6
requests==2.18.4
rq==0.10.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
